Update report graphql endpoint
