+++ server/graphql/Cargo.toml
@@ -26,6 +26,7 @@ graphql_stocktake = { path = "stocktake" }
 graphql_stocktake_line = { path = "stocktake_line" }
 graphql_batch_mutations = { path = "batch_mutations" }
 graphql_stock_line = { path = "stock_line" }
+graphql_form_schema = { path = "form_schema" }
 
 actix-web = { version = "4.0.1", default-features = false, features = [
     "macros",
