+++ server/graphql/form_schema/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "graphql_form_schema"
+version = "0.1.0"
+edition = "2018"
+
+[lib]
+path = "src/lib.rs"
+doctest = false
+
+[dependencies]
+
+repository = { path = "../../repository" }
+service = { path = "../../service" }
+util = { path = "../../util" }
+graphql_core = { path = "../core" }
+graphql_types = { path = "../types" }
+
+actix-web = { version = "4.0.1", default-features = false, features = ["macros"] }
+anymap = "0.12"
+async-graphql = { version = "3.0.35", features = ["dataloader", "chrono"] }
+async-graphql-actix-web = "3.0.35"
+async-trait = "0.1.30"
+chrono = { version = "0.4", features = ["serde"] }
+reqwest = { version = "0.11.10", features = ["json"] }
+serde = "1.0.126"
+serde_json = "1.0.66"
+thiserror = "1.0.30"
+anyhow = "1.0.56"
+
+[dev-dependencies]
+actix-rt = "2.6.0"
+assert-json-diff = "2.0.1"
+
+[features]
+default = ["sqlite"]
+sqlite = ["repository/sqlite"]
+postgres = ["repository/postgres"]
