+++ server/graphql/programs/src/mutations/mod.rs
@@ -2,7 +2,6 @@ pub mod allocate_number;
 pub mod delete_document;
 pub mod encounter;
 pub mod insert_document_registry;
-pub mod insert_form_schema;
 pub mod patient;
 pub mod program_enrolment;
 pub mod undelete_document;
