+++ server/graphql/programs/src/queries/mod.rs
@@ -2,8 +2,6 @@ pub mod document;
 pub use self::document::*;
 pub mod document_history;
 pub use self::document_history::*;
-pub mod json_schema;
-pub use self::json_schema::*;
 pub mod patient;
 pub use self::patient::*;
 pub mod patient_search;
