+++ server/graphql/programs/src/types/mod.rs
@@ -1,6 +1,5 @@
 pub(crate) mod document;
 pub(crate) mod document_registry;
 pub(crate) mod encounter;
-pub(crate) mod json_schema;
 pub(crate) mod program_enrolment;
 pub(crate) mod program_event;
