+++ server/graphql/types/src/types/json_schema.rs
